RAG/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.7
faiss-cpu==1.7.4
sentence-transformers==2.2.2
torch==2.1.1
transformers==4.35.2
ast-comments==1.1.0
tree-sitter==0.20.2
tree-sitter-python==0.20.4
gitpython==3.1.40
pydantic==2.5.0
redis==5.0.1
numpy==1.24.3
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.23.2
pre-commit==3.5.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1
